fix convert number, return float rather than int

+++ src/Shopmacher/IEProcessor/Csv/Helper/ConvertNumber.php
@@ -30,6 +30,6 @@ class ConvertNumber implements ConverterInterface
      */
     public function apply($value, $raw)
     {
-        return (int) $value;
+        return (float) $value;
     }
 }
